feat(extension): add background store with fade
